# CST-407-Application-Security-Foundations/coding/JWT/jwt-secured-api/src/main/java/edu/gcu/cst407/jwt/dto/LoginRequest.java
package edu.gcu.cst407.jwt.dto;

import jakarta.validation.constraints.NotBlank;

public class LoginRequest {

    @NotBlank
    private String username;

    @NotBlank
    private String password;

    public LoginRequest() {
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}