# swagger-spec.json
{"openapi":"3.0.1","info":{"title":"JWT Secured Orders API","description":"Practice API for CST-407 Activity 6 introducing JWT authentication","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/orders/{id}":{"get":{"tags":["orders-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["orders-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["orders-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/api/users/register":{"post":{"tags":["users-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/users/login":{"post":{"tags":["users-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/orders":{"get":{"tags":["orders-controller"],"operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderResponse"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["orders-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/users/me":{"get":{"tags":["users-controller"],"operationId":"currentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"OrderRequest":{"required":["description","orderDate","total"],"type":"object","properties":{"description":{"type":"string"},"total":{"minimum":0.0,"exclusiveMinimum":true,"type":"number"},"orderDate":{"type":"string","format":"date"}}},"OrderResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"description":{"type":"string"},"total":{"type":"number"},"orderDate":{"type":"string","format":"date"},"createdBy":{"type":"string"}}},"RegisterRequest":{"required":["fullName","password","username"],"type":"object","properties":{"username":{"maxLength":80,"minLength":4,"type":"string"},"password":{"maxLength":120,"minLength":8,"type":"string"},"fullName":{"maxLength":120,"minLength":2,"type":"string"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"fullName":{"type":"string"},"role":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"},"tokenType":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}